+++ NEWS.md
@@ -10,6 +10,7 @@
 * adds/exports `fact()` and `pseudo_id()`
 * adds `fixmes()` [#13](https://github.com/jmbarbone/jordan/issues/13)
 * adds `names_switch()` to switch names and values
+* adds an `invert` parameter to `complete_cases()` to filter for incomplete cases
 
 # jordan 0.1.2